feat: complete the forgot and reset page logic
