+++ src/pages/QuizForm/index.tsx
@@ -1,0 +1,8 @@
+
+import { Aside } from './styles';
+
+export function QuizForm() {
+  return (
+    <h1>QuizForm</h1>
+  );
+};
