+++ src/pages/QuizForm/styles.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const Aside = styled.aside``;
